requirements.txt
albumentations==0.4.5
ensemble-boxes==1.0.4
imageio==2.8.0
imgaug==0.2.6
llvmlite==0.33.0
matplotlib==3.2.2
numba==0.50.0
numpy==1.18.1
omegaconf==2.0.0
opencv-python==4.2.0.34
pandas==1.0.5
Pillow==5.4.1
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
timm==0.1.28
torch==1.4.0
torchvision==0.5.0
tqdm==4.46.1
